+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## 17.02.0
 - you now can sort or notes also alphabetically descending
   (for [Issue #504](https://github.com/pbek/QOwnNotes/issues/504))
+- disabled the `Copy` button in the Note editor font schema settings 
+  under macOS because of a Qt bug for Qt versions lower than 5.8
+  (for [Issue #503](https://github.com/pbek/QOwnNotes/issues/503))
 
 ## 17.01.14
 - you can now turn off the ownCloud / Nextcloud support in the settings to